@@ -16,60 +16,62 @@
     "lib/"
   ],
   "scripts": {
-    "build": "cross-env NODE_ENV=production node scripts/build.js",
+    "build": "cross-env NODE_ENV=production node ./scripts/build.js",
     "jest": "cross-env NODE_ENV=test jest \"$@\"",
     "lint": "yarn run lint-prettier && eslint . --cache",
-    "lint-prettier": "node scripts/prettier.js lint",
-    "prettier": "node scripts/prettier.js write",
-    "tck": "yarn run build && node packages/rsocket-tck/build/index.js \"$@\"",
-    "simple-cli": "yarn run build && node packages/rsocket-examples/build/SimpleCli.js \"$@\"",
-    "resume-example": "yarn run build && node packages/rsocket-examples/build/ResumeExample.js",
-    "lease-client-example": "yarn run build && node packages/rsocket-examples/build/LeaseClientExample.js \"$@\"",
+    "lint-prettier": "node ./scripts/prettier.js lint",
+    "prettier": "node ./scripts/prettier.js write",
+    "tck": "yarn run build && node ./packages/rsocket-tck/build/index.js \"$@\"",
+    "simple-cli": "yarn run build && node ./packages/rsocket-examples/build/SimpleCli.js \"$@\"",
+    "resume-example": "yarn run build && node ./packages/rsocket-examples/build/ResumeExample.js",
+    "lease-client-example": "yarn run build && node ./packages/rsocket-examples/build/LeaseClientExample.js \"$@\"",
     "test": "f() { EXIT=0; yarn run typecheck || EXIT=$?; yarn run jest \"$@\" || EXIT=$?; exit $EXIT; }; f",
-    "typecheck": "flow check packages/"
+    "typecheck": "flow check ./packages/"
   },
   "workspaces": [
     "packages/*"
   ],
   "devDependencies": {
-    "babel-cli": "^6.9.0",
-    "babel-core": "^6.9.0",
-    "babel-eslint": "^7.2.1",
-    "babel-plugin-minify-replace": "^0.4.3",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.24.1",
-    "babel-plugin-transform-flow-strip-types": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-plugin-transform-runtime": "^6.9.0",
-    "babel-polyfill": "^6.9.0",
-    "babel-preset-fbjs": "^2.0.0",
-    "babel-runtime": "^6.6.1",
-    "buffer": "^5.0.6",
-    "chalk": "^1.1.3",
-    "cross-env": "^6.0.3",
-    "eslint": "^4.18.2",
-    "eslint-config-fb-strict": "^19.0.1",
-    "eslint-plugin-babel": "^4.1.1",
-    "eslint-plugin-flowtype": "^2.30.3",
-    "eslint-plugin-jasmine": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prefer-object-spread": "^1.1.0",
-    "eslint-plugin-react": "^6.10.3",
-    "eslint-plugin-relay": "^1.3.12",
-    "fbjs": "^0.8.1",
-    "fbjs-scripts": "^0.7.1",
-    "flow-bin": "^0.78.0",
-    "glob": "^7.1.1",
-    "jest": "^19.0.2",
-    "lerna": "^2.9.0",
-    "object-assign": "^4.1.0",
-    "prettier": "0.22.0",
-    "rollup": "^0.41.6",
-    "rollup-plugin-babel": "^2.7.1",
-    "rollup-plugin-commonjs": "^8.0.2",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "ws": "^2.3.1"
+    "@babel/cli": "^7.12.8",
+    "@babel/core": "^7.12.9",
+    "babel-jest": "^26.6.3",
+    "babel-eslint": "^10.1.0",
+    "babel-plugin-minify-replace": "^0.5.0",
+    "@babel/plugin-transform-async-to-generator": "^7.12.1",
+    "@babel/plugin-proposal-async-generator-functions": "^7.12.1",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/plugin-transform-modules-commonjs": "^7.12.1",
+    "@babel/plugin-transform-flow-strip-types": "^7.12.1",
+    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
+    "@babel/plugin-transform-runtime": "^7.12.1",
+    "@babel/polyfill": "^7.12.1",
+    "@babel/runtime": "^7.12.5",
+    "babel-preset-fbjs": "^3.3.0",
+    "chalk": "^4.1.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^7.15.0",
+    "eslint-config-fb-strict": "^26.0.0",
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-flowtype": "^5.2.0",
+    "eslint-plugin-jasmine": "^4.1.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prefer-object-spread": "^1.2.1",
+    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-relay": "^1.8.1",
+    "eslint-plugin-jest": "^24.1.3",
+    "fbjs": "^3.0.0",
+    "fbjs-scripts": "^3.0.0",
+    "flow-bin": "^0.139.0",
+    "glob": "^7.1.6",
+    "jest": "^26.6.3",
+    "lerna": "^3.22.1",
+    "object-assign": "^4.1.1",
+    "prettier": "2.2.1",
+    "rollup": "^2.34.2",
+    "@rollup/plugin-babel": "^5.2.2",
+    "@rollup/plugin-commonjs": "^17.0.0",
+    "@rollup/plugin-node-resolve": "^11.0.0",
+    "ws": "^7.4.1"
   },
   "devEngines": {
     "node": ">=4.x"
