@@ -15,7 +15,7 @@ export class CompositeMetadata {
   constructor(buffer: Buffer) {
     this._buffer = buffer;
   }
-
+  // $FlowFixMe
   [Symbol.iterator]() {
     return entriesIterator(this._buffer);
   }
@@ -23,99 +23,104 @@ export class CompositeMetadata {
 
 /**
  * Encode a new sub-metadata information into a composite metadata {@link CompositeByteBuf
-  * buffer}, without checking if the {@link String} can be matched with a well known compressable
-  * mime type. Prefer using this method and {@link #encodeAndAddMetadata(CompositeByteBuf,
-  * ByteBufAllocator, WellKnownMimeType, ByteBuf)} if you know in advance whether or not the mime
-  * is well known. Otherwise use {@link #encodeAndAddMetadataWithCompression(CompositeByteBuf,
-  * ByteBufAllocator, String, ByteBuf)}
-  *
-  * @param compositeMetaData the buffer that will hold all composite metadata information.
-  * @param allocator the {@link ByteBufAllocator} to use to create intermediate buffers as needed.
-  * @param customMimeType the custom mime type to encode.
-  * @param metadata the metadata value to encode.
-  */
+ * buffer}, without checking if the {@link String} can be matched with a well known compressable
+ * mime type. Prefer using this method and {@link #encodeAndAddMetadata(CompositeByteBuf,
+ * ByteBufAllocator, WellKnownMimeType, ByteBuf)} if you know in advance whether or not the mime
+ * is well known. Otherwise use {@link #encodeAndAddMetadataWithCompression(CompositeByteBuf,
+ * ByteBufAllocator, String, ByteBuf)}
+ *
+ * @param compositeMetaData the buffer that will hold all composite metadata information.
+ * @param allocator the {@link ByteBufAllocator} to use to create intermediate buffers as needed.
+ * @param customMimeType the custom mime type to encode.
+ * @param metadata the metadata value to encode.
+ */
 // see #encodeMetadataHeader(ByteBufAllocator, String, int)
 export function encodeAndAddCustomMetadata(
   compositeMetaData: Buffer,
   customMimeType: string,
   metadata: Buffer,
 ): Buffer {
-  return Buffer.concat([
-    compositeMetaData,
-    encodeCustomMetadataHeader(customMimeType, metadata.byteLength),
-    metadata,
-  ]);
+  return Buffer.concat(
+    ([
+      compositeMetaData,
+      encodeCustomMetadataHeader(customMimeType, metadata.byteLength),
+      metadata,
+    ]: Buffer[]),
+  );
 }
 
 /**
  * Encode a new sub-metadata information into a composite metadata {@link CompositeByteBuf
-  * buffer}.
-  *
-  * @param compositeMetaData the buffer that will hold all composite metadata information.
-  * @param allocator the {@link ByteBufAllocator} to use to create intermediate buffers as needed.
-  * @param knownMimeType the {@link WellKnownMimeType} to encode.
-  * @param metadata the metadata value to encode.
-  */
+ * buffer}.
+ *
+ * @param compositeMetaData the buffer that will hold all composite metadata information.
+ * @param allocator the {@link ByteBufAllocator} to use to create intermediate buffers as needed.
+ * @param knownMimeType the {@link WellKnownMimeType} to encode.
+ * @param metadata the metadata value to encode.
+ */
 // see #encodeMetadataHeader(ByteBufAllocator, byte, int)
 export function encodeAndAddWellKnownMetadata(
   compositeMetaData: Buffer,
-  knownMimeType: WellKnownMimeType | Number,
+  knownMimeType: WellKnownMimeType | number,
   metadata: Buffer,
 ): Buffer {
   let mimeTypeId: number;
 
   if (Number.isInteger(knownMimeType)) {
-    mimeTypeId = (knownMimeType: number);
+    mimeTypeId = ((knownMimeType: any): number);
   } else {
-    mimeTypeId = (knownMimeType: WellKnownMimeType).identifier;
+    mimeTypeId = ((knownMimeType: any): WellKnownMimeType).identifier;
   }
 
-  return Buffer.concat([
-    compositeMetaData,
-    encodeWellKnownMetadataHeader(mimeTypeId, metadata.byteLength),
-    metadata,
-  ]);
+  return Buffer.concat(
+    ([
+      compositeMetaData,
+      encodeWellKnownMetadataHeader(mimeTypeId, metadata.byteLength),
+      metadata,
+    ]: Buffer[]),
+  );
 }
 
 /**
-  * Decode the next metadata entry (a mime header + content pair of {@link ByteBuf}) from a {@link
-  * ByteBuf} that contains at least enough bytes for one more such entry. These buffers are
-  * actually slices of the full metadata buffer, and this method doesn't move the full metadata
-  * buffer's {@link ByteBuf#readerIndex()}. As such, it requires the user to provide an {@code
-  * index} to read from. The next index is computed by calling {@link #computeNextEntryIndex(int,
-  * ByteBuf, ByteBuf)}. Size of the first buffer (the "header buffer") drives which decoding method
-  * should be further applied to it.
-  *
-  * <p>The header buffer is either:
-  *
-  * <ul>
-  *   <li>made up of a single byte: this represents an encoded mime id, which can be further
-  *       decoded using {@link #decodeMimeIdFromMimeBuffer(ByteBuf)}
-  *   <li>made up of 2 or more bytes: this represents an encoded mime String + its length, which
-  *       can be further decoded using {@link #decodeMimeTypeFromMimeBuffer(ByteBuf)}. Note the
-  *       encoded length, in the first byte, is skipped by this decoding method because the
-  *       remaining length of the buffer is that of the mime string.
-  * </ul>
-  *
-  * @param compositeMetadata the source {@link ByteBuf} that originally contains one or more
-  *     metadata entries
-  * @param entryIndex the {@link ByteBuf#readerIndex()} to start decoding from. original reader
-  *     index is kept on the source buffer
-  * @param retainSlices should produced metadata entry buffers {@link ByteBuf#slice() slices} be
-  *     {@link ByteBuf#retainedSlice() retained}?
-  * @return a {@link ByteBuf} array of length 2 containing the mime header buffer
-  *     <strong>slice</strong> and the content buffer <strong>slice</strong>, or one of the
-  *     zero-length error constant arrays
-  */
+ * Decode the next metadata entry (a mime header + content pair of {@link ByteBuf}) from a {@link
+ * ByteBuf} that contains at least enough bytes for one more such entry. These buffers are
+ * actually slices of the full metadata buffer, and this method doesn't move the full metadata
+ * buffer's {@link ByteBuf#readerIndex()}. As such, it requires the user to provide an {@code
+ * index} to read from. The next index is computed by calling {@link #computeNextEntryIndex(int,
+ * ByteBuf, ByteBuf)}. Size of the first buffer (the "header buffer") drives which decoding method
+ * should be further applied to it.
+ *
+ * <p>The header buffer is either:
+ *
+ * <ul>
+ *   <li>made up of a single byte: this represents an encoded mime id, which can be further
+ *       decoded using {@link #decodeMimeIdFromMimeBuffer(ByteBuf)}
+ *   <li>made up of 2 or more bytes: this represents an encoded mime String + its length, which
+ *       can be further decoded using {@link #decodeMimeTypeFromMimeBuffer(ByteBuf)}. Note the
+ *       encoded length, in the first byte, is skipped by this decoding method because the
+ *       remaining length of the buffer is that of the mime string.
+ * </ul>
+ *
+ * @param compositeMetadata the source {@link ByteBuf} that originally contains one or more
+ *     metadata entries
+ * @param entryIndex the {@link ByteBuf#readerIndex()} to start decoding from. original reader
+ *     index is kept on the source buffer
+ * @param retainSlices should produced metadata entry buffers {@link ByteBuf#slice() slices} be
+ *     {@link ByteBuf#retainedSlice() retained}?
+ * @return a {@link ByteBuf} array of length 2 containing the mime header buffer
+ *     <strong>slice</strong> and the content buffer <strong>slice</strong>, or one of the
+ *     zero-length error constant arrays
+ */
 export function decodeMimeAndContentBuffersSlices(
   compositeMetadata: Buffer,
   entryIndex: number,
 ): Buffer[] {
-  const mimeIdOrLength: number = compositeMetadata.readInt8(entryIndex, 1);
+  const mimeIdOrLength: number = compositeMetadata.readInt8(entryIndex);
   let mime: Buffer;
   let toSkip = entryIndex;
   if (
-    (mimeIdOrLength & STREAM_METADATA_KNOWN_MASK) === STREAM_METADATA_KNOWN_MASK
+    (mimeIdOrLength & STREAM_METADATA_KNOWN_MASK) ===
+    STREAM_METADATA_KNOWN_MASK
   ) {
     mime = compositeMetadata.slice(toSkip, toSkip + 1);
     toSkip += 1;
@@ -166,20 +171,22 @@ export function decodeMimeAndContentBuffersSlices(
  * properly contains such a mime type.
  *
  * <p>The buffer must at least have two readable bytes, which distinguishes it from the {@link
-  * #decodeMimeIdFromMimeBuffer(ByteBuf) compressed id} case. The first byte is a size and the
-  * remaining bytes must correspond to the {@link CharSequence}, encoded fully in US_ASCII. As a
-  * result, the first byte can simply be skipped, and the remaining of the buffer be decoded to the
-  * mime type.
-  *
-  * <p>If the mime header buffer is less than 2 bytes long, returns {@code null}.
-  *
-  * @param flyweightMimeBuffer the mime header {@link ByteBuf} that contains length + custom mime
-  *     type
-  * @return the decoded custom mime type, as a {@link CharSequence}, or null if the input is
-  *     invalid
-  * @see #decodeMimeIdFromMimeBuffer(ByteBuf)
-  */
-export function decodeMimeTypeFromMimeBuffer(flyweightMimeBuffer: Buffer) {
+ * #decodeMimeIdFromMimeBuffer(ByteBuf) compressed id} case. The first byte is a size and the
+ * remaining bytes must correspond to the {@link CharSequence}, encoded fully in US_ASCII. As a
+ * result, the first byte can simply be skipped, and the remaining of the buffer be decoded to the
+ * mime type.
+ *
+ * <p>If the mime header buffer is less than 2 bytes long, returns {@code null}.
+ *
+ * @param flyweightMimeBuffer the mime header {@link ByteBuf} that contains length + custom mime
+ *     type
+ * @return the decoded custom mime type, as a {@link CharSequence}, or null if the input is
+ *     invalid
+ * @see #decodeMimeIdFromMimeBuffer(ByteBuf)
+ */
+export function decodeMimeTypeFromMimeBuffer(
+  flyweightMimeBuffer: Buffer,
+): string {
   if (flyweightMimeBuffer.length < 2) {
     throw new Error('Unable to decode explicit MIME type');
   }
@@ -192,7 +199,7 @@ export function decodeMimeTypeFromMimeBuffer(flyweightMimeBuffer: Buffer) {
 export function encodeCustomMetadataHeader(
   customMime: string,
   metadataLength: number,
-) {
+): Buffer {
   const metadataHeader: Buffer = createBuffer(4 + customMime.length);
   // reserve 1 byte for the customMime length
   // /!\ careful not to read that first byte, which is random at this point
@@ -235,7 +242,7 @@ export function encodeCustomMetadataHeader(
 export function encodeWellKnownMetadataHeader(
   mimeType: number,
   metadataLength: number,
-) {
+): Buffer {
   const buffer: Buffer = Buffer.alloc(4);
 
   buffer.writeUInt8(mimeType | STREAM_METADATA_KNOWN_MASK);
@@ -244,7 +251,7 @@ export function encodeWellKnownMetadataHeader(
   return buffer;
 }
 
-function* entriesIterator(buffer: Buffer) {
+function* entriesIterator(buffer: Buffer): Generator<Entry, void, any> {
   const length = buffer.byteLength;
   let entryIndex = 0;
 
@@ -286,10 +293,10 @@ export interface Entry {
   +content: Buffer,
 
   /**
- * Returns the MIME type of the entry, if it can be decoded.
- *
- * @return the MIME type of the entry, if it can be decoded, otherwise {@code null}.
- */
+   * Returns the MIME type of the entry, if it can be decoded.
+   *
+   * @return the MIME type of the entry, if it can be decoded, otherwise {@code null}.
+   */
   +mimeType: ?string,
 }
 
@@ -302,11 +309,11 @@ export class ExplicitMimeTimeEntry implements Entry {
     this._type = type;
   }
 
-  get content() {
+  get content(): Buffer {
     return this._content;
   }
 
-  get mimeType() {
+  get mimeType(): string {
     return this._type;
   }
 }
@@ -320,7 +327,7 @@ export class ReservedMimeTypeEntry implements Entry {
     this._type = type;
   }
 
-  get content() {
+  get content(): Buffer {
     return this._content;
   }
 
@@ -328,7 +335,7 @@ export class ReservedMimeTypeEntry implements Entry {
    * {@inheritDoc} Since this entry represents a compressed id that couldn't be decoded, this is
    * always {@code null}.
    */
-  get mimeType() {
+  get mimeType(): ?string {
     return undefined;
   }
 
@@ -338,7 +345,7 @@ export class ReservedMimeTypeEntry implements Entry {
    *
    * @return the reserved, but unknown {@link WellKnownMimeType} for this entry
    */
-  get type() {
+  get type(): number {
     return this._type;
   }
 }
@@ -379,12 +386,12 @@ export class WellKnownMimeTypeEntry implements Entry {
  * STREAM_METADATA_KNOWN_MASK) == STREAM_METADATA_KNOWN_MASK}).
  *
  * <p>If there is no readable byte, the negative identifier of {@link
-  * WellKnownMimeType#UNPARSEABLE_MIME_TYPE} is returned.
-  *
-  * @param mimeBuffer the buffer that should next contain the compressed mime id byte
-  * @return the compressed mime id, between 0 and 127, or a negative id if the input is invalid
-  * @see #decodeMimeTypeFromMimeBuffer(ByteBuf)
-  */
+ * WellKnownMimeType#UNPARSEABLE_MIME_TYPE} is returned.
+ *
+ * @param mimeBuffer the buffer that should next contain the compressed mime id byte
+ * @return the compressed mime id, between 0 and 127, or a negative id if the input is invalid
+ * @see #decodeMimeTypeFromMimeBuffer(ByteBuf)
+ */
 function decodeMimeIdFromMimeBuffer(mimeBuffer: Buffer): number {
   if (!isWellKnownMimeType(mimeBuffer)) {
     return UNPARSEABLE_MIME_TYPE.identifier;
@@ -397,10 +404,12 @@ function computeNextEntryIndex(
   headerSlice: Buffer,
   contentSlice: Buffer,
 ): number {
-  return currentEntryIndex +
+  return (
+    currentEntryIndex +
     headerSlice.byteLength + // this includes the mime length byte
     3 + // 3 bytes of the content length, which are excluded from the slice
-    contentSlice.byteLength;
+    contentSlice.byteLength
+  );
 }
 
 function isWellKnownMimeType(header: Buffer): boolean {
@@ -410,7 +419,7 @@ function isWellKnownMimeType(header: Buffer): boolean {
 const STREAM_METADATA_KNOWN_MASK = 0x80; // 1000 0000
 const STREAM_METADATA_LENGTH_MASK = 0x7f; // 0111 1111
 
-function isAscii(buffer: Buffer, offset: number) {
+function isAscii(buffer: Buffer, offset: number): boolean {
   let isAscii = true;
   for (let i = offset, length = buffer.length; i < length; i++) {
     if (buffer[i] > 127) {
