@@ -24,6 +24,6 @@ const PACKAGES_DIR = path.resolve(__dirname, '../packages');
 module.exports = function getPackages() {
   return fs
     .readdirSync(PACKAGES_DIR)
-    .map(file => path.resolve(PACKAGES_DIR, file))
-    .filter(f => fs.lstatSync(path.resolve(f)).isDirectory());
+    .map((file) => path.resolve(PACKAGES_DIR, file))
+    .filter((f) => fs.lstatSync(path.resolve(f)).isDirectory());
 };
