@@ -14,7 +14,7 @@
  */
 'use strict';
 
-const babel = require('babel-core');
+const babel = require('@babel/core');
 const createCacheKeyFunction = require('fbjs-scripts/jest/createCacheKeyFunction');
 const moduleMap = require('fbjs/module-map');
 const path = require('path');
@@ -36,7 +36,7 @@ const babelOptions = {
 };
 
 module.exports = {
-  process: function(src, filename) {
+  process: function (src, filename) {
     const options = Object.assign({}, babelOptions, {
       filename: filename,
     });
